+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/DashMediaPeriod.java
@@ -54,6 +54,7 @@ import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.source.chunk.ChunkSampleStream;
 import androidx.media3.exoplayer.source.chunk.ChunkSampleStream.EmbeddedSampleStream;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
+import androidx.media3.exoplayer.trackselection.TrackSelection;
 import androidx.media3.exoplayer.upstream.Allocator;
 import androidx.media3.exoplayer.upstream.CmcdConfiguration;
 import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
@@ -121,6 +122,7 @@ import java.util.regex.Pattern;
   private boolean canReportInitialDiscontinuity;
   private long initialStartTimeUs;
   private long endPositionUs;
+  private boolean readingSuppressedWaitingForInitialDiscontinuity;
 
   public DashMediaPeriod(
       int id,
@@ -319,6 +321,9 @@ import java.util.regex.Pattern;
     if (canReportInitialDiscontinuity) {
       canReportInitialDiscontinuity = false;
       initialStartTimeUs = positionUs;
+      if (mayHaveAnyStreamWithPendingInitialDiscontinuity()) {
+        setSuppressReadOnAllStreams(true);
+      }
     }
     return positionUs;
   }
@@ -343,11 +348,7 @@ import java.util.regex.Pattern;
 
   @Override
   public boolean continueLoading(LoadingInfo loadingInfo) {
-    boolean result = compositeSequenceableLoader.continueLoading(loadingInfo);
-    if (hasAnyStreamWithPendingInitialDiscontinuity()) {
-      setSuppressReadOnAllStreams(true);
-    }
-    return result;
+    return compositeSequenceableLoader.continueLoading(loadingInfo);
   }
 
   @Override
@@ -362,11 +363,12 @@ import java.util.regex.Pattern;
 
   @Override
   public long readDiscontinuity() {
-    for (ChunkSampleStream<DashChunkSource> sampleStream : sampleStreams) {
-      if (sampleStream.consumeInitialDiscontinuity()) {
-        if (!hasAnyStreamWithPendingInitialDiscontinuity()) {
-          setSuppressReadOnAllStreams(false);
-        }
+    if (readingSuppressedWaitingForInitialDiscontinuity) {
+      boolean hasDiscontinuity = tryConsumeInitialDiscontinuityFromStreams();
+      if (!mayHaveAnyStreamWithPendingInitialDiscontinuity()) {
+        setSuppressReadOnAllStreams(false);
+      }
+      if (hasDiscontinuity) {
         return initialStartTimeUs;
       }
     }
@@ -417,9 +419,9 @@ import java.util.regex.Pattern;
 
   // Internal methods.
 
-  private boolean hasAnyStreamWithPendingInitialDiscontinuity() {
+  private boolean mayHaveAnyStreamWithPendingInitialDiscontinuity() {
     for (ChunkSampleStream<DashChunkSource> sampleStream : sampleStreams) {
-      if (sampleStream.hasInitialDiscontinuity()) {
+      if (sampleStream.mayHaveInitialDiscontinuity()) {
         return true;
       }
     }
@@ -427,9 +429,18 @@ import java.util.regex.Pattern;
   }
 
   private void setSuppressReadOnAllStreams(boolean suppressRead) {
+    readingSuppressedWaitingForInitialDiscontinuity = suppressRead;
     for (ChunkSampleStream<DashChunkSource> sampleStream : sampleStreams) {
       sampleStream.setSuppressRead(suppressRead);
     }
+  }
+
+  private boolean tryConsumeInitialDiscontinuityFromStreams() {
+    boolean consumedDiscontinuity = false;
+    for (ChunkSampleStream<DashChunkSource> sampleStream : sampleStreams) {
+      consumedDiscontinuity |= sampleStream.consumeInitialDiscontinuity();
+    }
+    return consumedDiscontinuity;
   }
 
   private int[] getStreamIndexToTrackGroupIndex(ExoTrackSelection[] selections) {
@@ -875,11 +886,17 @@ import java.util.regex.Pattern;
       embeddedClosedCaptionTrackFormats.add(embeddedTrackFormats[embeddedTrackCount]);
       embeddedTrackCount++;
     }
-
     PlayerTrackEmsgHandler trackPlayerEmsgHandler =
         manifest.dynamic && enableEventMessageTrack
             ? playerEmsgHandler.newPlayerTrackEmsgHandler()
             : null;
+    long firstChunkStartTimeUs =
+        getFirstChunkStartTimeUs(
+            positionUs, manifest, periodIndex, trackGroupInfo.adaptationSetIndices);
+    boolean handleInitialDiscontinuity =
+        canReportInitialDiscontinuity
+            && !areAllSamplesSyncSamples(
+                manifest, periodIndex, trackGroupInfo.adaptationSetIndices, selection);
     DashChunkSource chunkSource =
         chunkSourceFactory.createDashChunkSource(
             manifestLoaderErrorThrower,
@@ -909,7 +926,8 @@ import java.util.regex.Pattern;
             drmEventDispatcher,
             loadErrorHandlingPolicy,
             mediaSourceEventDispatcher,
-            canReportInitialDiscontinuity,
+            handleInitialDiscontinuity,
+            firstChunkStartTimeUs,
             downloadExecutorSupplier != null ? downloadExecutorSupplier.get() : null);
     stream.setEndPositionUs(endPositionUs);
     synchronized (this) {
@@ -917,6 +935,46 @@ import java.util.regex.Pattern;
       trackEmsgHandlerBySampleStream.put(stream, trackPlayerEmsgHandler);
     }
     return stream;
+  }
+
+  private static long getFirstChunkStartTimeUs(
+      long positionsUs, DashManifest manifest, int periodIndex, int[] adaptationSetIndices) {
+    // Use the first representation as all representations in one adaptation set (or in grouped
+    // adaptation sets) need to be segment-aligned.
+    Representation representation =
+        manifest
+            .getPeriod(periodIndex)
+            .adaptationSets
+            .get(adaptationSetIndices[0])
+            .representations
+            .get(0);
+    @Nullable DashSegmentIndex segmentIndex = representation.getIndex();
+    if (segmentIndex == null) {
+      return C.TIME_UNSET;
+    }
+    long periodDurationUs = manifest.getPeriodDurationUs(periodIndex);
+    return segmentIndex.getTimeUs(segmentIndex.getSegmentNum(positionsUs, periodDurationUs));
+  }
+
+  private static boolean areAllSamplesSyncSamples(
+      DashManifest manifest,
+      int periodIndex,
+      int[] adaptationSetIndices,
+      TrackSelection trackSelection) {
+    List<AdaptationSet> adaptationSets = manifest.getPeriod(periodIndex).adaptationSets;
+    ImmutableList.Builder<Representation> representationsBuilder = ImmutableList.builder();
+    for (int adaptationSetIndex : adaptationSetIndices) {
+      representationsBuilder.addAll(adaptationSets.get(adaptationSetIndex).representations);
+    }
+    ImmutableList<Representation> representations = representationsBuilder.build();
+    for (int i = 0; i < trackSelection.length(); i++) {
+      Representation representation = representations.get(trackSelection.getIndexInTrackGroup(i));
+      if (!MimeTypes.allSamplesAreSyncSamples(
+          representation.format.sampleMimeType, representation.format.codecs)) {
+        return false;
+      }
+    }
+    return true;
   }
 
   @Nullable